Solve some import dependency cycles that were exposed by absolute import
transitfeed has many import cycles, which is a bad thing. Converting
to absolute import exposed some of them. This changes adds the minimal
changes required to make imports work, for example by moving some
types from `problems.py` to `errors.py`.
